django_blog/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.1.1
django-taggit==6.0.0
django-widget-tweaks==1.5.0
idna==3.7
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
